+++ program/op.go
@@ -1,8 +1,8 @@
 package program
 
-type Op func(intcode *Intcode, index int, op int64, os Os) int
+type Op func(intcode *Intcode, index int64, op int64, os Os) int64
 
-var StdInstructions = map[int]Op{
+var StdInstructions = map[int64]Op{
 	1:  Add,
 	2:  Mul,
 	3:  Input,
@@ -15,7 +15,7 @@ var StdInstructions = map[int]Op{
 	99: Halt,
 }
 
-func Add(intcode *Intcode, index int, op int64, os Os) int {
+func Add(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 3)
 
 	a := intcode.Get(mode[0], index+1)
@@ -25,7 +25,7 @@ func Add(intcode *Intcode, index int, op int64, os Os) int {
 	return index + 4
 }
 
-func Mul(intcode *Intcode, index int, op int64, os Os) int {
+func Mul(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 3)
 
 	a := intcode.Get(mode[0], index+1)
@@ -35,7 +35,7 @@ func Mul(intcode *Intcode, index int, op int64, os Os) int {
 	return index + 4
 }
 
-func Input(intcode *Intcode, index int, op int64, os Os) int {
+func Input(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 1)
 
 	a := os.Read()
@@ -44,7 +44,7 @@ func Input(intcode *Intcode, index int, op int64, os Os) int {
 	return index + 2
 }
 
-func Output(intcode *Intcode, index int, op int64, os Os) int {
+func Output(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 1)
 
 	a := intcode.Get(mode[0], index+1)
@@ -53,33 +53,33 @@ func Output(intcode *Intcode, index int, op int64, os Os) int {
 	return index + 2
 }
 
-func JumpIfTrue(intcode *Intcode, index int, op int64, os Os) int {
+func JumpIfTrue(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 2)
 
 	a := intcode.Get(mode[0], index+1)
 	b := intcode.Get(mode[1], index+2)
 
 	if a != 0 {
-		return int(b)
+		return b
 	}
 
 	return index + 3
 }
 
-func JumpIfFalse(intcode *Intcode, index int, op int64, os Os) int {
+func JumpIfFalse(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 2)
 
 	a := intcode.Get(mode[0], index+1)
 	b := intcode.Get(mode[1], index+2)
 
 	if a == 0 {
-		return int(b)
+		return b
 	}
 
 	return index + 3
 }
 
-func LessThan(intcode *Intcode, index int, op int64, os Os) int {
+func LessThan(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 3)
 
 	a := intcode.Get(mode[0], index+1)
@@ -94,7 +94,7 @@ func LessThan(intcode *Intcode, index int, op int64, os Os) int {
 	return index + 4
 }
 
-func Equals(intcode *Intcode, index int, op int64, os Os) int {
+func Equals(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 3)
 
 	a := intcode.Get(mode[0], index+1)
@@ -109,15 +109,15 @@ func Equals(intcode *Intcode, index int, op int64, os Os) int {
 	return index + 4
 }
 
-func AdjustBase(intcode *Intcode, index int, op int64, os Os) int {
+func AdjustBase(intcode *Intcode, index int64, op int64, os Os) int64 {
 	mode := modes(op, 1)
 
 	a := intcode.Get(mode[0], index+1)
-	intcode.AdjustBase(int(a))
+	intcode.AdjustBase(a)
 
 	return index + 2
 }
 
-func Halt(intcode *Intcode, index int, op int64, os Os) int {
+func Halt(intcode *Intcode, index int64, op int64, os Os) int64 {
 	return -1
 }
